refactor (all) : major refactor to be modular compatible
