+++ core/src/main/java/core/Application.java
@@ -9,7 +9,7 @@
  * This class is based on the AlbUtil project.
  * 
  */
-package main;
+package core;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
