+++ database/src/main/java/es/uniovi/asw/database/App.java
@@ -1,0 +1,13 @@
+package es.uniovi.asw.database;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
