+++ database/src/main/resources/application.properties
@@ -1,0 +1,5 @@
+spring.data.mongodb.database=aswdb
+spring.data.mongodb.host=ds237445.mlab.com
+spring.data.mongodb.port=37445
+spring.data.mongodb.username=app
+spring.data.mongodb.password=1234
