+++ domain/src/main/java/org/domain/App.java
@@ -1,0 +1,13 @@
+package org.domain;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
