+++ rest/src/main/java/view/ParticipantsController.java
