+++ rest/src/main/java/view/ParticipantsDataController.java
