+++ rest/src/main/java/view/UserNotFoundException.java
