+++ rest/src/main/resources/master.csv
@@ -1,0 +1,3 @@
+3,Sensor
+2,Entity
+1,Person
