+++ rest/src/main/resources/templates/data.html
@@ -1,0 +1,102 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+
+<head>
+  <meta charset="UTF-8" />
+  <!-- to use the latest rendering engine in case it doesn't-->
+  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+
+  <!-- Adjusts the viewport and the scale of the browser -->
+  <meta name="viewport" content="width = device-width, initial-scale = 1" />
+
+  <!-- Bootstrap CDN in order to have support for Bootstrap's CSS and js -->
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous" />
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
+  <link rel="stylesheet" type="text/css" href="https://s3-eu-west-1.amazonaws.com/cssfonts/SF/sf-font.css" />
+  <link rel="stylesheet" href="signin.css" />
+  <title>Login</title>
+
+
+</head>
+
+<body>
+
+  <!-- container class in order to tell BS to treat it as a container and it is usually container -->
+  <div class="container-fluid well span6">
+
+    <div class="col-md-6">
+      <h1><strong>Account information</strong></h1>
+      <div class="table-responsive">
+        <table class="table table-condensed table-responsive table-user-information">
+          <tbody>
+            <tr>
+              <td>
+                <strong>
+                          Name
+                     </strong>
+              </td>
+              <td class="text-primary">
+                <span th:text="${name}" />
+              </td>
+            </tr>
+            <tr>
+              <td>
+                <strong>
+                          Id
+                     </strong>
+              </td>
+              <td class="text-primary">
+                <span th:text="${id}" />
+              </td>
+            </tr>
+            <tr>
+              <td>
+                <strong>
+                         Location
+                     </strong>
+              </td>
+              <td class="text-primary">
+                <span th:text="${location}" />
+              </td>
+            </tr>
+
+            <tr>
+              <td>
+                <strong>
+                         Email
+                     </strong>
+              </td>
+              <td class="text-primary">
+                <span th:text="${email}" />
+              </td>
+            </tr>
+            <tr>
+              <td>
+                <strong>
+                         Kind Of Agent
+                     </strong>
+              </td>
+              <td class="text-primary">
+                <span th:text="${kind}" />
+              </td>
+            </tr>
+            <tr><td></td><td></td></tr>
+          </tbody>
+        </table>
+      </div>
+    </div>
+    <!-- <a class="btn btn-primary glyphicon glyphicon-cog " href="/passMenu">Change password</a> -->
+  </div>
+
+
+  <!-- jQuery (necessary for Bootstrap's JavaScript plugins) -->
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js"></script>
+
+
+  <!-- in order to include tether for BS 4 -->
+  <script src="https://npmcdn.com/tether@1.2.4/dist/js/tether.min.js"></script>
+  <!-- Include all compiled plugins (below), or include individual files as needed -->
+  <script src="bootstrap.min.js"></script>
+</body>
+
+</html>
